Added methods to fix leading and trailing tones, so we can recognize short silences between notes and display high-pitch anomalies next to the notes they're part of. The latter seems to mostly work, but could use improvement.

+++ channelmodel.cpp
@@ -39,7 +39,7 @@ QVariant ChannelModel::data(const QModelIndex &index, int role) const
             return QVariant(tone.duty);
         break;
         case LengthRole:
-            return QVariant(tone.length);
+            return QVariant((qint64)tone.length);
         break;
         case NameRole:
             return QVariant(tone.name());
